+++ lib/src/all_units.dart
@@ -1,6 +1,7 @@
 import 'database/units/imperials/74_z_speeder_bike.dart' as _i;
 import 'database/units/imperials/at_st.dart' as _i;
 import 'database/units/imperials/boba_fett.dart' as _i;
+import 'database/units/imperials/bossk.dart' as _i;
 import 'database/units/imperials/darth_vader.dart' as _i;
 import 'database/units/imperials/e_web_heavy_blaster_team.dart' as _i;
 import 'database/units/imperials/emperor_palpatine.dart' as _i;
@@ -22,7 +23,9 @@ import 'database/units/rebels/luke_skywalker.dart' as _i;
 import 'database/units/rebels/rebel_commandos.dart' as _i;
 import 'database/units/rebels/rebel_commandos_strike_team.dart' as _i;
 import 'database/units/rebels/rebel_officer.dart' as _i;
+import 'database/units/rebels/rebel_pathfinders.dart' as _i;
 import 'database/units/rebels/rebel_troopers.dart' as _i;
+import 'database/units/rebels/sabrine_wren.dart' as _i;
 import 'database/units/rebels/t_47_airspeeder.dart' as _i;
 import 'database/units/rebels/wookie_warriors.dart' as _i;
 
@@ -30,6 +33,7 @@ final aggregate = [
   _i.$74ZSpeederBikes,
   _i.atSt,
   _i.bobaFett,
+  _i.bossk,
   _i.darthVader,
   _i.eWebHeavyBlasterTeam,
   _i.emperorPalpatine,
@@ -51,7 +55,9 @@ final aggregate = [
   _i.rebelCommandos,
   _i.rebelCommandosStrikeTeam,
   _i.rebelOfficer,
+  _i.rebelPathfinders,
   _i.rebelTroopers,
+  _i.sabrineWren,
   _i.t47Airspeeder,
   _i.wookieWarriors
 ];
@@ -61,6 +67,7 @@ class Units {
   static final $74ZSpeederBikes = _i.$74ZSpeederBikes;
   static final atSt = _i.atSt;
   static final bobaFett = _i.bobaFett;
+  static final bossk = _i.bossk;
   static final darthVader = _i.darthVader;
   static final eWebHeavyBlasterTeam = _i.eWebHeavyBlasterTeam;
   static final emperorPalpatine = _i.emperorPalpatine;
@@ -82,7 +89,9 @@ class Units {
   static final rebelCommandos = _i.rebelCommandos;
   static final rebelCommandosStrikeTeam = _i.rebelCommandosStrikeTeam;
   static final rebelOfficer = _i.rebelOfficer;
+  static final rebelPathfinders = _i.rebelPathfinders;
   static final rebelTroopers = _i.rebelTroopers;
+  static final sabrineWren = _i.sabrineWren;
   static final t47Airspeeder = _i.t47Airspeeder;
   static final wookieWarriors = _i.wookieWarriors;
 }