+++ lib/src/database/upgrades/blaster/the_darksaber.dart
@@ -1,0 +1,21 @@
+import 'package:swlegion/catalog.dart';
+import 'package:swlegion/swlegion.dart';
+
+final theDarksaber = Upgrade(
+  id: 'the-darksaber',
+  name: 'The Darksaber',
+  type: UpgradeSlot.blaster,
+  restrictedToUnit: [
+    Units.sabrineWren.toRef(),
+  ],
+  weapon: Weapons.theDarksaber,
+  points: 25,
+  keywordsForUnit: {
+    UnitKeyword.kDauntless: '',
+    UnitKeyword.kImmunePierce: 'Melee',
+  },
+  waves: const [
+    Wave.wSabineWrenOperativeExpansion,
+  ],
+  isUnique: true,
+);