+++ lib/src/database/weapons/upgrades/a_300_long_range_config.dart
@@ -1,0 +1,10 @@
+import 'package:swlegion/swlegion.dart';
+
+final a300LongRangeBlasterRifle = Weapon(
+  name: 'A-300 Long-Range Blaster Rifle',
+  minRange: 1,
+  maxRange: 4,
+  dice: {
+    AttackDice.white: 1,
+  },
+);